Major restructuring
I've reworked most of the compiler in order to make it more
maintainable and less complex.

The new compiler has a text based IR which allows for a
modular architecture. This will allow the user to plug
in their own passes easily and choose the ordering of
existing passes etc.

Currently, there is no codegen for x86-64 implemented but
this is just a foundation for the compiler going forward.

Plans are to implement another IR in three-address-code form
for backends to consume which should ease register allocation
and lowering to assembly.

- New modular architecture
- Simplified implementation
- Stack based IR with textual format