Renamed core words to cp, mv & rm and added stdlib.klx.
Renamed the core words to try and avoid too many
naming conflicts but also to describe their intent
better.

Added a very basic stdlib file which gives access to some
common kind of arithmetic and stack manipulation words.

Updated the syntax highlighting file in accordance
with the above and also added a new highlighting file
for the klaxon IR format (KIR).

Renamed the compile.sh script to klx and we also run
the m4 preprocessor on the source file before passing
it to klaxon to enable use of include and macros.

Fixed an issue when parsing type annotations.
Previously, annotations which had no out values would
cause the compiler to generate an error that it expected
an identifier but returning no values is valid.

Added a dedicated locale string for type annotation errors.

Removed extraneous space being printed in the KIR serialiser.