Add indirect branch elimination pass
Added a new optimisation pass to collapse indirect
jumps to a block which then unconditionally jumps
to another block.

This is very useful for heavily nested if-else chains.

Give up when trying to do constant folding beyond
a function call. The problem with trying to fold
beyond the bounds of a call is that the function
call itself may produce some values that are only
known at compile time in which case there is nothing
for the constant folder to reduce at compile time.

Removed any notion of arg and out instructions.