Major restructuring of optimisation passes
Function inlining and dead code elimination were previously
broken and the code was awkward to work with due to having
to try and preserve consistency in the same buffer.

Three issues have been addressed in this commit:
1. Functioning inlining now renumbers blocks correctly
2. Dead code elimination now only retains functions with "main" as an ancestor
3. Iterators to the IR are now stable due to the use of a double buffer approach

Function inlining was broken previously due to not renumbering blocks
after they were inlined. This would mean that multiple calls to the
same function which had been inlined woudl result in duplicate
blocks which would break the control flow of the program.

The new inliner also doesn't count block/def/end/ret instructions.

Dead code elimination was previously broken due to preserving
functions which were called but not by a common ancestor
("main" in this case). All it would take to preserve a function
was to call it _anywhere_ in the program even if the parent function
of that call was itself dead.

We now use a double buffering like approach to optimisation passes.
The original IR is passed in and supposed to remain unchanged while
the output IR is supposed to be mutated and will become the next
input buffer for the next pass. This gives us some rather nice
properties like stability of reference which makes inlining in
particular very easy.

Constant folding and indirect branch elimination have yet to be
moved over to the new architecture but should be fairly easy.