Simplified lexing greatly and avoid hard-coding strings
Using a simpler architecture for the lexer which allows us
to specify token strings in a single place and have it work
everywhere. This is in contrast to the previous lexer which
required updating multiple unrelated pieces of the code in
order to change tokens.

Switch from shorthand names `cp`, `mv` and `rm` to
`copy`, `move` and `remove`.

Fixed issue where the IR parser would only accept non-keyword
identifiers for user defined functions. This meant functions
named "block" for example would result in a parsing error.

"copy", "move" and "remove" are now considered proper keywords
and as such cannot be used as identifiers.

Removed any hard-coding of strings in calls to error functions
and instead look up the appropriate string representation of
a token instead.