Switch to register/stack based hybrid IR
Large refactoring of the codebase to follow the new IR spec.

A register based IR should allow us to lower the high level
code to something more akin to machine code easily.

The core syntax now allows for external and internal
declarations aswell as allowing for inline declaration
in definitions.

- Updated syntax highlighting for kakoune
- Updated EBNF for IR and language